# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Monitoring ST2023",
        "version": "1.0"
    },
    "paths": {
...
        "/api/ruta": {
            "get": {
                "tags": [
                    "ruta"
                ],
                "summary": "Get List",
                "description": "-",
                "operationId": "ruta/index",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "current page to show",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "kode wilayah and pagination",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "per_page": {
                                        "type": "integer"
                                    },
                                    "kode_wilayah": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "return array model ruta"
                    }
                }
            },
            "post": {
                "tags": [
                    "ruta"
                ],
                "summary": "Store ruta",
                "description": "-",
                "operationId": "ruta/store",
                "requestBody": {
                    "description": "form ruta",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "kode_prov",
                                    "kode_kab",
                                    "kode_kec",
                                    "kode_desa",
                                    "id_sls",
                                    "id_sub_sls",
                                    "nurt",
                                    "kepala_ruta",
                                    "start_time",
                                    "end_time",
                                    "start_latitude",
                                    "end_latitude",
                                    "start_longitude",
                                    "end_longitude"
                                ],
                                "properties": {
                                    "kode_prov": {
                                        "type": "string"
                                    },
                                    "kode_kab": {
                                        "type": "string"
                                    },
                                    "kode_kec": {
                                        "type": "string"
                                    },
                                    "kode_desa": {
                                        "type": "string"
                                    },
                                    "id_sls": {
                                        "type": "string"
                                    },
                                    "id_sub_sls": {
                                        "type": "string"
                                    },
                                    "nurt": {
                                        "type": "integer"
                                    },
                                    "kepala_ruta": {
                                        "type": "string"
                                    },
                                    "start_time": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "end_time": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "start_latitude": {
                                        "type": "decimal"
                                    },
                                    "end_latitude": {
                                        "type": "decimal"
                                    },
                                    "start_longitude": {
                                        "type": "decimal"
                                    },
                                    "end_longitude": {
                                        "type": "decimal"
                                    },
                                    "subsektor1_a": {
                                        "type": "integer",
                                        "maximum": 1,
                                        "minimum": 0
                                    },
                                    "subsektor1_b": {
                                        "type": "integer",
                                        "maximum": 1,
                                        "minimum": 0
                                    },
                                    "subsektor2_a": {
                                        "type": "integer",
                                        "maximum": 1,
                                        "minimum": 0
                                    },
                                    "subsektor2_b": {
                                        "type": "integer",
                                        "maximum": 1,
                                        "minimum": 0
                                    },
                                    "subsektor3_a": {
                                        "type": "integer",
                                        "maximum": 1,
                                        "minimum": 0
                                    },
                                    "subsektor3_b": {
                                        "type": "integer",
                                        "maximum": 1,
                                        "minimum": 0
                                    },
                                    "subsektor4_a": {
                                        "type": "integer",
                                        "maximum": 1,
                                        "minimum": 0
                                    },
                                    "subsektor4_b": {
                                        "type": "integer",
                                        "maximum": 1,
                                        "minimum": 0
                                    },
                                    "subsektor4_c": {
                                        "type": "integer",
                                        "maximum": 1,
                                        "minimum": 0
                                    },
                                    "subsektor5_a": {
                                        "type": "integer",
                                        "maximum": 1,
                                        "minimum": 0
                                    },
                                    "subsektor5_b": {
                                        "type": "integer",
                                        "maximum": 1,
                                        "minimum": 0
                                    },
                                    "subsektor5_c": {
                                        "type": "integer",
                                        "maximum": 1,
                                        "minimum": 0
                                    },
                                    "subsektor6_a": {
                                        "type": "integer",
                                        "maximum": 1,
                                        "minimum": 0
                                    },
                                    "subsektor6_b": {
                                        "type": "integer",
                                        "maximum": 1,
                                        "minimum": 0
                                    },
                                    "subsektor6_c": {
                                        "type": "integer",
                                        "maximum": 1,
                                        "minimum": 0
                                    },
                                    "subsektor7_a": {
                                        "type": "integer",
                                        "maximum": 1,
                                        "minimum": 0
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
        "/api/ruta/many": {
            "post": {
                "tags": [
                    "ruta"
                ],
                "summary": "Store Many Ruta",
                "description": "You can update array here by put the enc ID on array object. If not, just set ID with '' value",
                "operationId": "ruta/many",
                "parameters": [
                    {
                        "name": "Bearer Token",
                        "in": "header",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "in data, just send object with property kode_prov, kode_kab, kode_kec, kode_desa, id_sls, id_sub_sls, nurt, kepala_ruta,\n     *          start_time, end_time, start_latitude, end_latitude, start_longitude, end_longitude, subsektor1_a, subsektor1_b, subsektor1_c, subsektor2_a,\n     *          subsektor2_b,subsektor3_a,subsektor3_b,subsektor4_a,subsektor4_b,subsektor4_c,subsektor5_a,subsektor5_b,subsektor5_c,subsektor6_a,\n     *          subsektor6_b,subsektor6_c,subsektor7_a",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "data"
                                ],
                                "properties": {
                                    "data": {
                                        "type": "array",
                                        "items": {
                                            "description": "data",
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
...
    }
}